Config / Extract jade renderer
