+++ config/wesh/fl_breadcrumb.ts
@@ -1,4 +1,4 @@
-import {fl_element, fl_container, fl_jadeRenderer, fl_text} from './fl_comp'
+import {fl_element, fl_container, fl_text} from './fl_comp'
 import * as fl_c from './fl_common'
 import * as fl_m from './fl_manager_comp'
