+++ config/wesh/main.ts
@@ -5,7 +5,8 @@ import * as fl_servers from './fl_servers'
 import * as fl_platforms from './fl_platforms'
 import * as fl_clients from './fl_clients'
 import * as fl_events from './fl_events'
-import {fl_container, fl_element, fl_jadeRenderer} from './fl_comp'
+import {fl_container, fl_element} from './fl_comp'
+import {fl_jadeRenderer} from './fl_comp_jade';
 
 interface String {
   repeat(count: number): string
